moved main validator to separate file
